Add a very blunt attempt to normalize the content entities
This commit contains an approach that will normalize all the entity
inside of the entity normalizer. What we want to do next is to have
different normalizers to take care of each sub object. For instance
we want to have separate normalizes to normalize a field item list
and another separated one to normalize a field item, and the another
one to normalize the property.